Use a unique control ID for the site icon control when used in the AMP Templates panel context.
If a non-unique control ID is used, the control is effectively _moved_ from one section to another, meaning that trying to use the same control (by ID) in multiple sections results in the site icon control not being present in the core "Site Identity" section any longer. A unique control ID paired with the 'settings' option fixes this problem.