Add installation/run instructions to comments in bash script.
Add the run-amp-validator.js node script.

This currently is able to run the amp-validator command with one test URL.  The commented code parses our XML data file to gather the URLs of our published posts so we can process them through the validator.

TO DO NEXT: be able to run the gathered URLs through the validator without overloading the resources.