Fix timing error that still existed and add control pages in the bin folder to ensure testing is working as expected for this tool.
Needed to return the nested call to the validator promise to prevent misfiring of our supposedly async Promises.  This was discovered by included a success and failure control html page in addition to the successful WPTest pages.  These HTML files are currently included for further testing of the tool, but should probably be removed once the tool is stable.

This version also copies the readFromReadable() and readFromUrl() functions from amp-html/validator/index.js so that we don't need to modify their file to allow for the export of the new module.  Not sure if this is the better solution, since we would then have to update these functions whenever the amp-html project updates them, but this might be a workable solution if there is a delay in or reason not to include the export line here: https://github.com/iamchrissmith/amphtml/blob/validator/readFromUrlExports/validator/index.js#L123