Store Test Data Locally:
Add wptest.xml data to Git and add an empty amptest.xml file for future amp-specific content tests.
Add a second question to the install script to ask the user if they want to install this data.
Add command to the cURL command for wptest.xml so that it tests if our local file is out of date and if it is replace our local version.

Move all test files into the /tests/assets/ folder for better organization.