AMP: Integrated FasterImage, removed extraction from meta data
Refactored Image Sanitizer and Dimension Extractor to use FasterImage to get dimensions for all images that need them concurrently.
As a result, it can now fetch dimensions for 100 images in 2 seconds whereas it previously took about 40 seconds. It takes less than a
second to get the dimensions for 10 remote images, and .002 seconds to get their dimensions when they are stored in transients (without memcached).
Removed the extract_from_attachment_metadata method for getting dimensions because it had to run multiple database queries
(triggered by attachment_url_to_post_id() and wp_get_attachment_metadata) for every image every time an amp page was viewed. With the new low "cost"
of extracting dimensions from the images themselves, it did not seem to warrant retaining the additional complexity.