Preserve inline styles: remove action hook from sanitizer
Instead of hooking into `amp_post_template_css` directly from the
sanitizer, this creates a generic `get_styles` method in the Base
sanitizer class, which is then used to pass the styles as a string
back to `AMP_Post_Template`, where it can be output in a post
template action.

This creates the possibility for other sanitizers to generate
and output dynamic styles as well.