Sanitize audio/video: update source attributes
Both the audio and video sanitizers use the same routine to iterate
through the source elements within the main node and filter/sanitize
their attributes. However, the filtered attributes are not added back to
the source element before it is appended to the new amp-audio or
amp-video element. This is problematic when the attribute filtering
method converts URLs from http to https, but the converted URL does not
make it back into the source element.

This commit addresses the issue by updating each source element's
attributes before it is appended.