Replace explode with preg_split.
In cases where Jetpack's `safecss_filter_attr` is used in place of core's function (and possibly other scenarios), there may be semi-colons in unexpected places.

For example, `background-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...);`

In that case, `explode` will break the style, splitting it in the middle of a value. With `preg_split` we can specify semi-colons that are not inside parentheses.