Allow disabling AMP Customizer later
Previously, you had to hook very early on `plugins_loaded` to be able to
disable the Customizer. This was not possible for sites that only have
access to the theme. This switches around the load order of the
Customizer so that the filter to disable the AMP Customizer can run any
time before `after_setup_theme`.