804 : Remove the requirement that 'data-multi-size' be empty.
This file was generated from a build script,
In the repo https://github.com/ampproject/amphtml
The spec file seems to require an empty value:
extensions/amp-ad/validator-amp-ad.protoascii
But examples of <amp-ad> seem to have the value.
Like <amp-ad data-multi-size=320x50.
@see examples/a4a-multisize.amp.html.
So remove the line in the PHP file that requires it to be empty.