804 : Add allowed protocols to sanitization file.
Before, there was an issue in the file that created:
class-amp-allowed-tags-generated.php.
amphtml-update.sh had a conditional:
if isinstance(value, list)
But the allowed_protocol attribute has a type of:
google.protobuf.internal.containers.RepeatedScalarFieldContainer
So that conditional failed.
And allowed_protocol wasn't added.
So instead, make that an 'else' statement.
That change only resulted in 'allowed_tags' being added.