Issue 848: Allow theme support of canonical AMP.
If themes register this with add_theme_support( 'amp' ),
the plugin will be in 'canonical mode.'
It won't display the 'rel' link,
And the theme will still use its own template.
But if any other argumnents are passed to add_theme_support( 'amp',
It will retain 'paired mode.'
Add unit tests for this new feature.