Issue #843: Validate Gutenberg blocks for AMP compliance.
In wp-admin/post.php, load a script.
This overwrites the edit() function of registered blocks.
The new function validates the block's content.
And conditionally displays a notice.
@todo: consider not using this for all blocks.
Some blocks are valid once the back-end sanitizer alters them.
Like the 'Audio' block.
This simply uses the save() function of the block to get the markup.