Display the URLs at the bottom of the CPT post.php page.
The <ul> in print_validation_errors_meta_box()
is only indented, with almost no change.
Its class is moved to the new parent <div>.
If the recheck reveals no errors,
it redirects to the CPT edit.php page.
And displays 'The rechecked URL has no validation error.'
If it still has an error, it stays on post.php,
with a message stating that the error remains.
@todo: consider a scenario where there is
more than 1 URL.
On rechecking and finding no error,
it might make sense to redirect to redirect
to that same post.php page.
The user might want to then recheck all of the URLs.
It's likely that the error won't exist there.