Store the block validation errors, avoiding lookup in every edit().
Before, this got the block errors from getCurrentPost(),
in every block's edit() method.
Instead, store these in the module,
every time there is an action.
It would be better to use a reducer
for a specific action.
But so far, I haven't found a way to do that
for REQUEST_POST_UPDATE_SUCCESS.