Use the new blockAttrs to find a match with errors.
This can match blocks that add a stripped script.
Like the 'Categories' block.
That wasn't part of the content,
so it wasn't found before.
Also, refactor existing logic into:
doNameAndAttributesMatch().
This is better in a function,
as it uses a for loop to return a value.
Otherwise, it would need to set a flag.