Remove the conditional around the add_action for preview scripts.
There's a check for is_customize_preview() around add_action().
But the callback dequeue_customize_preview_scripts()
has its own check for is_customize_preview_iframe().
And that has a check for is_customize_preview().
So this ensures that the body of the callback
never runs.
Maybe I misread the intent of this, though.
This also adds unit tests.