Set default layout of <amp-iframe> and <amp-video> to intrinsic.
Inspired by Mike Cranteas earlier commit,
which applied to <amp-img>
Before, this set the layout of those elements to responsive if the height and width weren't empty.
This does the same thing, only with 'intrinsic.'
Initial testing on Native AMP and paired mode
shows that this displays as expected.
According to the documentation,
'This layout works very well for most AMP elements, including amp-img, amp-video,'
@see https://www.ampproject.org/docs/design/responsive/control_layout#what-if-the-layout-attribute-isn%E2%80%99t-specified?