Fall back to Site Icon if Custom Logo exceeds dimensions.
It's very possible for the Custom Logo to be
more than the 600 x 60px required:
https://developers.google.com/search/docs/data-types/article#logo-guidelines
If so, fall back to the site icon as the publisher logo.
When the 32 x 32px Site Icon was the logo,
it didn't cause an error in the
Google Structured Data Testing tool.
Even though it didn't follow the guidelines.