Display schema.org image data for 'attachment' post type.
As reported on a support topic, on Native AMP,
on 'attachment' pages, there's no schema.org
data for the image.
So add a conditional to check for this post type.
It's possible that the attachment is a video,
like a .mov file.
In that case, it won't return an image.
wp_get_attachment_image_src() will return false.
@see https://wordpress.org/support/topic/structured-data-error-for-attachment-pages-media/