Keep the 'mode' of 'disabled' (Classic).
This might not be the right solution.
Before, there was an issue when selecting
the 'Template Mode' of classic.
It saved properly, but it didn't appear
on the front end.
It looks like it was overriden on
line 143 of AMP_Options_Menu::render_theme_support().
This fixes the issue locally,
but I'm not sure if it changes
what the code was intended for.
Or if it has unwanted side-effects.