Reset last validation errors memory when blocks are out of sync.
Blocks get out of sync in different scenarios.  For example, on page load, the block uids will change as the editor loads up and blocks are registered.

This commit compares the last known block order to the current block order.  If not equal, then it resets the last saved validation errors memory.  In doing so, if there are validation errors, then `handleValidationErrorsStateChange()` will continue processing instead of bailing out.