Update `uid` to `clientId`.
Gutenberg is removing `uid` and replacing it everywhere with `clientId`.  The change is included in 3.4.0 and `uid` will be removed in 3.5.0.

@see https://github.com/WordPress/gutenberg/blob/41fd116ca1f73a9b5c962c66ea0c4aae338ea35e/docs/reference/deprecated.md#350

We see the impact of this change as `getBlock()` requires the `clientId` now instead of the `uid`.

This commit replaces all references to `uid` with `clientId`.  In doing so, it fixes the issue with block-level notifications disappearing when duplicating a block and then saving.