Address failed unit test
All of the created posts had a term of 'Uncategorizes' added.
Before, 'Uncategorized' wasn't countent in the term query,
because there were no posts.
So create a new term to add to all of the posts,
and count it.
So 'Uncategorized' won't be part of the term query,
as it has no posts associated with it.