Allow crawling templates the user has unchecked in 'Supported Templates'
In 'AMP Settings,' users can uncheck certain templates.
And this had blocked validation of them,
as is_amp_endpoint() returned false and they were redirected.
So if the user passes the flag to the WP-CLI command,
add a query var.
This forces all endpoints to be AMP endpoints, to allow validation.
Still, is_amp_endpoint might not be the right
place to make this change.