Validate author pages, including with include=is_author
Also, set $force_crawl_urls if there's an include argument.
You could pass a conditional to include that
has AMP disabled now.
Like if you use the command:
wp amp validate-site include=is_author
But 'Author' could be unchecked in
'AMP Settings' > 'Supported Templates'
In that case, this would need to force
validation of that URL,
as the URL would otherwise redirect to the non-AMP URL.