Refactor get_author_page_urls() to use round-robin validation
As Alberto and Weston suggested,
the site crawling could validate one URL
of each template at a time.
This prevents validating possibly thousands of
posts or pages before getting to other URLs.
So this accepts an $offset argument.
0 gets the first user, and 1 the second...
The optional second parameter sets a number to return.
So in the round-robin validation,
this could be 1.