Filter the term redirect location, to possibly add a query var
On clicking the 'Filter' button on the 'AMP Validation Errors' taxonomy page,
edit-tags.php processes the POST request that this submits.
Then, it redirects to a URL to display the page again.
This filter callback looks for a value for VALIDATION_ERROR_TYPE_QUERY_VAR in the  request.
That means that the user filtered by error type, like 'js_error'.
It then passes that value to the redirect URL as a query var,
So that the taxonomy page will be filtered for that error type.
@see AMP_Validation_Error_Taxonomy::add_error_type_clauses_filter() for the filtering of the 'where' clause, based on that query var.