Improve validation of children of invalid elements
Before, if there is an unaccepted invalid element like <details>,
its child elements are not validated.
Only on accepting the <details> validation error
is the <summary> validated.
So this adds elements to $fragment only if
the fragment will replace the node.
Also, this ensures that the original parent is reported.
Before, an invalid node's children replaced the invalid node,
and the childrens' parent was reassigned.
So this stores the parent in a new property.
This might not be the best strategy.