Make the default layout reponsive for the <amp-ooyala-player> block
As Claudio pointed out,
the default layout of fixed can cause it
to overflow the container.
responsive seems to keep the width to the container size
in Twenty Fifteen, Sixteen, and Seventeen.
Still, if the content area is floated,
like in a mobile-first grid, a fixed layout might work better.
Responsive layouts mean the element only has the width of its parent.
But there's a <select> in the block editor to change this.