Add a default height for the AMP Timeago block
As Claudio reported,
there's an error when using the AMP Timeago block
with the default values
The layout is inially responsive,
which requires both a width an height.
And another option of fixed also requires a width and height.
So set fixed-height as the default layout,
as set a default height of 20.
This is the value in the documentation example.
So at least this block with the default values
won't have a validation error

@see https://www.ampproject.org/ko/docs/reference/components/amp-timeago