Correct failed unit tests by allowing the 'amp' or '⚡' attribute in <html>
In the ammphtml repo's validator-main.protoascii,
the attrs object for html was removed.
@see https://github.com/ampproject/amphtml/commit/ee9ef4449122459dfb69951e1b103407c855bdb8#diff-695a0119f06368dee3ab53e1395ef725L77
It looks like the amphtml repo's validator has
a different way of validating the <html> tag now:
https://github.com/ampproject/amphtml/commit/ee9ef4449122459dfb69951e1b103407c855bdb8#diff-52a1cb778626c52da13d91b4849462d6R4746
So this adds an exclusion for <html amp> or <html ⚡>
in is_amp_allowed_attribute().
Before, this had returned false for those attributes,
and this resulted in the' am'pan ⚡' being removed.