Begin to address issue where a @font-face can be removed
Sometimes this appears in $parsed_selector['tags'],
and the logic look for the presence of the tag on the page.
This would apply well if it were an HTML tag like div,
but @font-face wouldn't appear as an element in an HTML document.
So add a conditional to check if the tag is in the
'allowed_at_rules'.
Still, this might not be the best solution.
Also, it looks like @font-face should appear at the top
of the stylesheet.