Remove the sizes attribute from thumbnail images in Twenty Nineteen
The AMP runtime sets an inline style on an <amp-img> based
on the sizes attribute if it's present.
For example, <amp-img style="width:calc(50vw)">.
Removing the 'sizes' attribute isn't ideal, but it looks
like it's not possible to override that inline style.
This is only a workaround until the issue is addressed
in the amphtml repo.