Fall back to the normal gallery shortcode markup for 'thumbnail' size.
If the 'gallery' shortcode has a 'size' attribute of 'thumbnail',
prevent outputting an <amp-carousel>.
That will often get thumbnail images around 150 x 150,
while the <amp-carousel> usually has a width of 600 and a height of 480.
That often means very low-resolution images.
So fall back to the normal 'gallery' shortcode callback, gallery_shortcode().