Prevent horizontall-compressed images in Gallery block
There's a style rule that prevents these from compressing.
img { object-fit: cover; }
This selector seems to be changed to amp-img in style sanitization,
but it doesn't have the same effect as when applied to img
So copy the style rule verbatim into a stylesheet.
This is in AMP_Gallery_Embed_Handler,
as it's too late to enqueue_style,
'wp_print_styles' has already run by the time
that sanitizer runs.