If there is an embed as an error source, display that instead of the theme
This might not be the right approach,
but this addresses an issue that Tonya found
in testing this with the Essence Pro (Genesis) theme.
Genesis has a function genesis_do_post_content()
that calls the_content().
So if an embed in the post content has a validation error,
the theme will be reported as a source of a validation error.
So this outputs the theme as a source only if
there's no embed source of the validation error.