Fix an issue with the featured image header being too wide
As Weston mentioned, this sometimes stretches to fill the entire screen width.
Part of the issue is that the AMP runtime sets an inline width
style based on the sizes attribute.
https://github.com/ampproject/amphtml/issues/17053
Also, there's a style rule in twentyseventeen/style.css
that changes the <amp-img> to display: block
The typical display for an <amp-img> is inline-block.
So override that rule, and ensure it's centered.