Override respsonsive embed styling in Gutenberg
There's a wrapper for some embed-type blocks,
like YouTube, Vimeo, and SoundCloud.
It wraps the block in a div.wp-block-embed__wrapper
And there's a style rule that gives this a padding-top.
In AMP, this adds a large gap at the top of the block.
So add a style rule that overrides this.