Ensure the second argument of date() is an int
The documentation of date() requires the 2nd argument to be an int
And the documentation of get_the_date() says it returns a string.
But locally, using PHP 7.2, it looked like get_the_date()
returns an int.
I couldn't reproduce this,
but it'll probably be good to ensure this is an int.