Fix an issue where <figure> can overflow its container
figure.wp-caption often has an inline width style value,
like <figure style="width: 1199px" ...>
This can cause it to overflow the container.
There used to be logic in AMP_Style_Sanitizer::filter_style()
that converted 'width' to 'max-width'.
Now that it's removed,
apply a style rule as a workaround.
This is in a stylesheet because this only applies to Classic mode.
This issue doesn't seem to exist in Paired or Native mode.