Move enqueue_block_editor_assets() into script and style methods
As Weston mentioned,
this breaks the method into separate methods.
Also, it keeps the STORY_CARD_CSS_SLUG registeration
on a 'wp_default_styles' callback.

It would have been good to run register_scripts()
on the 'wp_default_scripts' hook.
But get_current_screen() is null,
and that's needed to verify that it's
an AMP Story post type.