Fix an issue where the poster-square and poster-landscape images were wrong
As Pascal pointed out,
these were swapped.
These values come from the featured image.
But the featured image is often cropped down to 696 x 928.
So from that, it's not possible to get a 928 x 928 image, for example.
So instead, use the source image that was cropped, instead of the cropped image.
This is more likely to produce the right size image.