Slightly improve issue with the text block resized with background
As Miina found, when the Text block is resized small
and has a background width,
it appears too tall in the block editor.
This removes the padding-top and bottom,
leaving only padding-left and right.
It doesn't completely fix the difference in the
front-end and back-end display, but it looks better locally.