If there is no image width yet, don't display notice
Add a check for attachment.get( 'width' )
On uploading an image,
there is about a second where the attachment doesn't have a width,
in which time there's a notice:
The selected image is too small ( by pixels). It should have a size of at least 1200 by 675 pixels.
About a second later, it has a value,
and the notification doesn't appear anymore.