Begin the work to automaticall set poster images
This applies to the Video block and the
AMP Story block 'Background Media'.
So far, on selecting a video,
this looks to see if it has an image in the object.
If so, it uses that image as the poster.
Even this might need refinement, like seeing if that image is the proper size.
There's a lot more work needed on this PR, also.