Limit height declaration to only direct descendant
The error occurs because the `.textbox` further in uses `vertical-align: middle` to align itself relative vertically centered relative to its parent, but this only works if the text box doesn't inherit this 100% height.

Even with the 100% height declaration, the text box sometimes (in particular on first load) receives the proper less-than-100% height, but as soon as the outer block is resized, the text box would switch back to top-aligned regardless of fit.

With this change, the inner text box will always be vertically centered. This has been confirmed across Chrome, Firefox, Safari and MSEdge - all able to display the error before, all not showing the error after.