Rewritten block size tests
This commit rewrites how block size testing is performed in e2e testing for text blocks.

The old method (used inconsistently) looked at the `.wp-block.is-selected` directly and this worked for fitted blocks, but if any of the tests had tested non-fit blocks, they would have errored.

This new method looks at the actual resizable box inside the block and this works for both fit and non-fit text blocks.

A test has been added to verify that non-fit text blocks can be resized, but due to a bug (reported in #3199) non-fit text blocks can't actually be resized smaller. The test has been skipped for now, but will serve as regression test for fixing #3199.