Add accessibility label for video blocks
This fixes #3186 by adding a new attribute to AMP video blocks, `ampAriaLabel`, and giving the editor access to edit this variable in the "Video Settings" inspector panel.

Furthermore this attribute is pre-populated from the media object selected by looking at `alt_text` or `title` of existing media object, if either exists.

Finally, this value is used in the output on the AMP story by manipulating the output of code video blocks and attaching the `aria-label` attribute to the `<amp-video />` element if applicable.