Updated Status component to simple functional comp
This component serves the purpose of announcing to screen readers, that the search results have changed in the autocomplete. And it needs a small change to occur when the message of "query too short" is displayed, even as the user types more characters. Thus a number of commas were inserted, and this still happens based off query length.

However, I don't think this component works at all, as it only works if it was in the DOM on the very first load of the HTML. And it clearly wasn't for a plurality of reasons. But I'm okay with keeping it for a11y reasons even though it might not actually serve that purpose.