Use vfsStream to mock filesystem, thanks to Alain's suggestion
Instead of writing to the real filesystem,
this uses a mock one through vfsStream.
It adds filters so that get_template_directory()
points to the mock filesystem,
and it can add an amp/ directory there.
This could be used more widely,
but at the moment it only applies to this test method.