Add array_colum() pollyfill for PHP < 5.5
This code is copied from the [`ramsey/array_colum`](https://github.com/ramsey/array_column) package, to avoid adding a dependency for now. The plan is to move to PHP 5.6+ in the near future anyway, so this is just meant to quickly fix the issue until it can be dropped again after the version requirement bump.

All changes to the original code did not change the logic, they only adapted the code style to satisfy PHPCS.