Reimplement snap lines via React Context API
The old snaplines were very labor intensive for the browser. The main culprit turned out to be the storage medium itself, storing the data in the WP Data store. Just issuing `dispatch` takes 40ms - even when the state doesn't change at all.

This rewrite uses the React Context API in stead storing snap lines in state locally in the context for each page and passing the getters and setters to the relevant components via a context consumer accessible through a simple higher-order function.

A complete rewrite of both draggable and resizable to hook based functions is recommended, but outside the scope of this task.

Unit tests are still needed for the context functionality.