Fix resize for nonfit textbox
Reducing the size of nonfit textbox was broken - possibly due changes upstream. The fixed height of textbox is required for other section of the code to work correctly, however it broke reducing size here.

By setting the textbox height to `auto` before calculating `scrollHeight`, the true height can once again be established - original height willl be reset after change.

Furthermore, this commit also changes the `<=` to `<` in the limiting condition - to allow one-directional resize.

However there are still weird issues for rotated textboxes when hitting the minimum size, but that have probably existed for some time.