Add unit tests for new snapping context
As this seems to be the first unit test for a hook-using component, it required some fundamental changes to the setup to make sure only a single React installation is ever in the mix. The mock for `@wordpress/element` is mostly just copied from the original, but guarantees that the React version matches that of Jest's test runner.

If this is doable in a different way, please advise.

The actual test of the snapping context validates, that the correct snap lines are displayed under different circumstances.