Allow drag to different page
_Fixes #3311._

This WIP allows editors to drag elements to neighboring pages.

This is done by while dragging, detecting where the actual mouse cursor is relative to the current page. A highlight callback is invoked while dragging with the offset of the page currently under the cursor. If dropped above a neighboring page, a drop callback is invoked.

The callbacks then check if there even is a neigbor in this direction and then either highlight that page or clone the element to that page.

It doesn't work correctly for CTA's and Page Attachments yet.

The code is also a little weirdly structured with the bulk of the logic placed in `block-draggable`, which doesn't really sound like the right place - concern-wise. Perhaps a simple renaming solves it, but perhaps a larger restructuring is in order.