Generalised helper for checking if block is allowed on given page
Several locations checked if a block was allowed on a given page, including the media inserter, the shortcut handler, paste handler and now also multi-page drag handler.

All these instances have now been a) centralised and b) changed to use the built-in method of checking for allowed blocks inside another block using core block editor functionality.

This also removed the need for a specific test for the `ensureAllowedBlocksOnPaste` as it just used other functions now more specific and central to the issue.