Reorganised helper to aid in testing
In order to properly test `isBlockAllowedOnPage` there's a need to mock `select` responses. But mocking all of `@wordpress/data` while also using a ton of other wordpress packages resulted in a lot of problems.

Moving the function to its own file and only mocking the `select`'ed functions for this single use ise much cleaner and works perfectly.

Most of the helpers could probably benefit from being moved to their own files. It's a lot cleaner and the slight overhead of `select`'ing the same functions many times is a minor worry.