Updated logic to display non-allowed drop zones
This required restructuring some things, but also made the interface cleaner - `isBlockAllowedOnPage` is now not used inside the `getNeighborPageId`, which also was a little weird.

There's also an explicit call to reset hover markings. It feels a little dirty, but it works. It could be a separate callback for a cleaner interface, but I don't feel it's necessary here.

The highlighting of disallowed pages / drop zones for CTAs is done in CSS only by adding classes based on the legality of a drop on a given page combined with a class for the type of element currently being dragged. If any other future elements to be dragged will be disallowed on some pages, they same red-bordered highlight will be shown on those pages, but for now it's only applicable to CTAs. It could easily be restricted to CTAs only if the need arises in the future.