Changed to use `KeyboardShortcuts` handler
This changes the implementation from manual key code validation to using the `KeyboardShortcuts` component from `@wordpress/components`.

This however has the problem, that it doesn't work when you focus the elmeent with the mouse and press Shift+F10. If you tab to the element, it works fine, but if you focus with the mouse, it doesn't work. I don't know if that's a problem though.