Convert helper function to hook
This converts the helper function `isBlockAllowedOnPage` to the hook `useIsBlockAllowedOnPage` which ensures that selectors are properly updated once the registry changes and any component using it is properly re-rendered.

However, because some uses of `isBlockAllowedOnPage` is located in class components, the old function has only been deprecated for now, not completely removed.