Allow drag to different page (#3312)
This new feature allows editors to drag elements to neighboring pages.

This is done by while dragging, detecting where the actual mouse cursor is relative to the current page. A highlight callback is invoked while dragging with the offset of the page currently under the cursor. If dropped above a neighboring page, a drop callback is invoked.